Move a number of forward declarations in to the fwd.hpp files.
