--- nano/lib/stream_fwd.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <cstdint>
-#include <iosfwd>
-
-struct uint8_char_traits;
-namespace nano
-{
-using stream = std::basic_streambuf<uint8_t, uint8_char_traits>;
-}
