+++ nano/node/bandwidth_limiter.hpp
@@ -6,11 +6,6 @@
 
 namespace nano
 {
-class node_config;
-}
-
-namespace nano
-{
 class bandwidth_limiter_config final
 {
 public:
