+++ nano/node/bootstrap_ascending/account_sets.hpp
@@ -3,6 +3,7 @@
 #include <nano/lib/numbers.hpp>
 #include <nano/node/bootstrap/bootstrap_config.hpp>
 #include <nano/node/bootstrap_ascending/common.hpp>
+#include <nano/node/fwd.hpp>
 
 #include <boost/multi_index/hashed_index.hpp>
 #include <boost/multi_index/mem_fun.hpp>
@@ -15,12 +16,6 @@
 #include <random>
 
 namespace mi = boost::multi_index;
-
-namespace nano
-{
-class account_sets_config;
-class stats;
-}
 
 namespace nano
 {
