+++ nano/node/vote_cache.hpp
@@ -4,7 +4,9 @@
 #include <nano/lib/locks.hpp>
 #include <nano/lib/numbers.hpp>
 #include <nano/lib/utility.hpp>
+#include <nano/node/fwd.hpp>
 #include <nano/secure/common.hpp>
+#include <nano/secure/fwd.hpp>
 
 #include <boost/multi_index/hashed_index.hpp>
 #include <boost/multi_index/mem_fun.hpp>
@@ -21,17 +23,6 @@
 #include <vector>
 
 namespace mi = boost::multi_index;
-
-namespace nano
-{
-class node;
-class active_elections;
-class election;
-class stats;
-class vote;
-enum class vote_code;
-enum class vote_source;
-}
 
 namespace nano
 {
