+++ nano/secure/account_info.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <nano/lib/epoch.hpp>
+#include <nano/lib/fwd.hpp>
 #include <nano/lib/numbers.hpp>
-#include <nano/lib/stream_fwd.hpp>
 #include <nano/lib/timer.hpp>
 
 namespace nano
