+++ nano/secure/fwd.hpp
@@ -1,14 +1,17 @@
 #pragma once
 
-namespace nano::secure
-{
-class transaction;
-class write_transaction;
-class read_transaction;
-}
-
 namespace nano
 {
 class account_info;
+class ledger;
+class ledger_cache;
+class ledger_constants;
+class network_params;
 class vote;
+}
+namespace nano::secure
+{
+class read_transaction;
+class transaction;
+class write_transaction;
 }
