+++ nano/secure/vote.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <nano/lib/fwd.hpp>
 #include <nano/lib/numbers.hpp>
-#include <nano/lib/stream_fwd.hpp>
 #include <nano/lib/timer.hpp>
 #include <nano/lib/uniquer.hpp>
 
@@ -9,11 +9,6 @@
 #include <boost/property_tree/ptree_fwd.hpp>
 
 #include <vector>
-
-namespace nano
-{
-class object_stream;
-}
 
 namespace nano
 {
