+++ nano/store/fwd.hpp
@@ -1,9 +1,23 @@
 #pragma once
 
+namespace nano
+{
+enum class tables;
+}
 namespace nano::store
 {
+class account;
+class block;
 class component;
-class transaction;
-class write_transaction;
+class confirmation_height;
+class final_vote;
+class online_weight;
+class peer;
+class pending;
+class pruned;
 class read_transaction;
+class rep_weight;
+class transaction;
+class version;
+class write_transaction;
 }
